rename *Img to *Media, update deprecated parts, cleanup
