+++ packages/infoportal-client/src/features/Database/KoboEntry/DatabaseKoboAnswerView.tsx
@@ -15,7 +15,7 @@ import {useI18n} from '@/core/i18n'
 import {KoboMappedAnswer} from '@/core/sdk/server/kobo/KoboMapper'
 import {KoboSchemaHelper, NonNullableKey} from 'infoportal-common'
 import React, {useEffect, useMemo, useState} from 'react'
-import {KoboAttachedImg} from '@/shared/TableImg/KoboAttachedImg'
+import {KoboAttachedImg} from '@/shared/TableMedia/KoboAttachedMedia'
 import {Txt} from '@/shared/Txt'
 import {useKoboSchemaContext} from '@/features/KoboSchema/KoboSchemaContext'
 import {Datatable} from '@/shared/Datatable/Datatable'
