+++ packages/infoportal-client/src/features/Meal/Visit/MealVisitDashboard.tsx
@@ -13,7 +13,7 @@ import {MapSvgByOblast} from '@/shared/maps/MapSvgByOblast'
 import {ChartPieWidget} from '@/shared/charts/ChartPieWidget'
 import {IpBtn} from '@/shared/Btn'
 import {CommentsPanel} from '@/shared/CommentsPanel'
-import {KoboAttachedImg} from '@/shared/TableImg/KoboAttachedImg'
+import {KoboAttachedImg} from '@/shared/TableMedia/KoboAttachedMedia'
 import {KoboSubmissionFlat, KoboIndex, Meal_visitMonitoring, OblastIndex} from 'infoportal-common'
 import {NavLink} from 'react-router-dom'
 import {DataFilterLayout} from '@/shared/DataFilter/DataFilterLayout'
