+++ packages/infoportal-client/src/features/Meta/Table/MetaTable.tsx
@@ -13,7 +13,7 @@ import React, {useEffect, useMemo} from 'react'
 import {StateStatusIcon} from '@/shared/customInput/SelectStatus'
 import {useSession} from '@/core/Session/SessionContext'
 import {DatatableColumn} from '@/shared/Datatable/util/datatableType'
-import {TableImg} from '@/shared/TableImg/TableImg'
+import {TableImg} from '@/shared/TableMedia/TableImg'
 import Link from 'next/link'
 import {AppFeatureId} from '@/features/appFeatureId'
 import {databaseIndex} from '@/features/Database/databaseIndex'
