+++ packages/infoportal-client/src/features/Mpca/MpcaData/MpcaData.tsx
@@ -14,7 +14,7 @@ import {useAsync} from '@/shared/hook/useAsync'
 import {DatatableUtils} from '@/shared/Datatable/util/datatableUtils'
 import {Datatable} from '@/shared/Datatable/Datatable'
 import {DeduplicationStatusIcon} from '@/features/WfpDeduplication/WfpDeduplicationData'
-import {TableImg} from '@/shared/TableImg/TableImg'
+import {TableImg} from '@/shared/TableMedia/TableImg'
 import {StateStatusIcon} from '@/shared/customInput/SelectStatus'
 import {AppFeatureId} from '@/features/appFeatureId'
 import {databaseIndex} from '@/features/Database/databaseIndex'
