+++ packages/infoportal-client/src/features/Shelter/Data/ShelterTable.tsx
@@ -21,7 +21,7 @@ import {useI18n} from '@/core/i18n'
 import {Panel} from '@/shared/Panel'
 import {Box, useTheme} from '@mui/material'
 import {TableIcon, TableIconBtn} from '@/features/Mpca/MpcaData/TableIcon'
-import {KoboAttachedImg} from '@/shared/TableImg/KoboAttachedImg'
+import {KoboAttachedImg} from '@/shared/TableMedia/KoboAttachedMedia'
 import {Txt} from '@/shared/Txt'
 import {useShelterContext} from '@/features/Shelter/ShelterContext'
 import {IpInput} from '@/shared/Input/Input'
